requirements.txt
Flask==3.0.2
APScheduler==3.11.1
requests==2.31.0
python-dotenv==1.0.1
selenium==4.11.2
tweepy==4.14.0
beautifulsoup4==4.12.2
google-generativeai==0.3.2
feedparser==6.0.11
Flask-Limiter==3.5.1
gunicorn==21.2.0
pytest==8.0.0
pytest-cov==4.1.0
Flask-CORS==4.0.0
